requirements.txt
# RSS Collector Core Dependencies

# RSS and web fetching
feedparser>=6.0.10
httpx>=0.24.0
requests>=2.31.0

# Date and time handling
python-dateutil>=2.8.2
pytz>=2023.3

# Scheduling
APScheduler>=3.10.0
schedule>=1.2.0

# Configuration and data handling
# (json is built-in to Python)
# python-dotenv is used for environment variables
python-dotenv>=1.0.1

RSS Collector Core Dependencies
Development Dependencies
pytest==7.4.0
black==23.7.0
flake8==6.0.0
